// src/models/IoK8sApimachineryPkgApisMetaV1APIResource/types.ts
export interface IoK8sApimachineryPkgApisMetaV1APIResourceModelOptions {
  categories?: Array<string>;
  group?: string;
  kind: string;
  name: string;
  namespaced: boolean;
  shortNames?: Array<string>;
  singularName: string;
  storageVersionHash?: string;
  verbs: Array<string>;
  version?: string;
}