// src/models/IoK8sApimachineryPkgApisMetaV1APIResourceList/types.ts
import { IoK8sApimachineryPkgApisMetaV1APIResourceModelOptions as IoK8sApimachineryPkgApisMetaV1APIResource } from '../IoK8sApimachineryPkgApisMetaV1APIResource';

export interface IoK8sApimachineryPkgApisMetaV1APIResourceListModelOptions {
  apiVersion?: string;
  groupVersion: string;
  kind?: string;
  resources: Array<IoK8sApimachineryPkgApisMetaV1APIResource>;
}